login and nav bar logics
