+++ public/src/Pages/AdminMain.jsx
@@ -1,15 +1,17 @@
 import AdminCards from "../Components/AdminCards";
-import AdminNavBar from "../Components/AdminNavBar";
 import AdminOrders from "../Components/AdminOrders";
-import Footer from "../Components/Footer";
 
 const AdminMain = (props) => {
   return (
     <>
-      <AdminNavBar />
-      <AdminOrders orders={props.orders}/>
-      <AdminCards />
-      <Footer />
+      <AdminOrders
+        orders={props.orders}
+        update={props.update}
+      />
+      <AdminCards
+        products={props.products}
+        update={props.update}
+      />
     </>
   );
 };
